Structure change - move config objects into configuration
