+++ Source/Configuration/OAuthBearerTokensConfig.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-namespace Aksio.IngressMiddleware;
+
+namespace Aksio.IngressMiddleware.Configuration;
 
 /// <summary>
 /// Represents the configuration for OAuth bearer tokens.
