+++ Source/ITenantSourceIdentifierResolver.cs
@@ -1,6 +1,8 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Aksio.IngressMiddleware.Configuration;
+
 namespace Aksio.IngressMiddleware;
