+++ Source/Identity.cs
@@ -5,6 +5,7 @@ using System.Net;
 using System.Net.Http.Headers;
 using System.Text;
 using Aksio.Cratis.Execution;
+using Aksio.IngressMiddleware.Configuration;
 
 namespace Aksio.IngressMiddleware;
