+++ Source/OAuthBearerTokens.cs
@@ -6,6 +6,7 @@ using System.Net;
 using System.Text;
 using System.Text.Json;
 using Aksio.Cratis.Execution;
+using Aksio.IngressMiddleware.Configuration;
 using Aksio.IngressMiddleware.Security;
 using Microsoft.IdentityModel.Tokens;
