+++ Source/Program.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Aksio.IngressMiddleware;
+using Aksio.IngressMiddleware.Configuration;
 
 UnhandledExceptionsManager.Setup();
