+++ Source/RouteSourceIdentifierResolver.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Text.RegularExpressions;
+using Aksio.IngressMiddleware.Configuration;
 
 namespace Aksio.IngressMiddleware;
