+++ Source/Security/ClaimsSourceIdentifierResolver.cs
@@ -3,6 +3,7 @@
 
 using System.Text;
 using System.Text.Json.Nodes;
+using Aksio.IngressMiddleware.Configuration;
 
 namespace Aksio.IngressMiddleware.Security;
